% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3funnel.R
\name{d3funnel}
\alias{d3funnel}
\title{d3funnel}
\usage{
d3funnel(data, ..., width = 300, height = 300)
}
\arguments{
\item{data}{should be a data.frame with two columns, 
the first column will be used as label(character), and the second column(numeric) is used as value}
}
\description{
d3funnel
}

